+++ packages/proto-gen-api-client/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2](https://github.com/planjs/stan/compare/proto-gen-api-client@0.2.0...proto-gen-api-client@0.3.2) (2021-02-28)
+
+
+### Bug Fixes
+
+* deps ([0a9442d](https://github.com/planjs/stan/commit/0a9442da1156886b299fe7b3c8234c61e7143066))
+
+
+
+
+
 ## [0.3.1](https://github.com/planjs/stan/compare/proto-gen-api-client@0.3.0...proto-gen-api-client@0.3.1) (2021-02-27)
 
 **Note:** Version bump only for package proto-gen-api-client