+++ packages/stan-env/core/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/planjs/stan/compare/stan-env-core@0.1.0...stan-env-core@0.2.2) (2021-02-28)
+
+
+### Bug Fixes
+
+* deps ([0a9442d](https://github.com/planjs/stan/commit/0a9442da1156886b299fe7b3c8234c61e7143066))
+* local deps ([05c103d](https://github.com/planjs/stan/commit/05c103ded88889e9d4b1d5c8b79e1d3cf043b559))
+
+
+
+
+
 ## [0.2.1](https://github.com/planjs/stan/compare/stan-env-core@0.2.0...stan-env-core@0.2.1) (2021-02-27)